Change file structure changed signup routes models and controller to users
