+++ backend/src/controllers/users.controller.js
@@ -1,19 +1,60 @@
-import { getUserModel } from "../models/users.model.js";
+import * as UserModel from "../models/users.model.js";
 
-export const getUser = async (req, res) => {
-    try {
-      const userId = req.params.id;
-      console.log(`getUser called with id: ${userId}`);
-      const result = await getUserModel(userId);
-      console.log('Query result:', result.recordset);
-      if (result.recordset.length === 0) {
-        res.status(404).json({ error: 'User not found' });
-      } else {
-        res.json(result.recordset[0]);
-      }
-    } catch (err) {
-      console.error('Error fetching user data:', err);
-      res.status(500).json({ error: 'Failed to retrieve user data' });
-    }
-  };
+// Get all users
+export const getAllUsers = async (req, res) => {
+  try {
+    const users = await UserModel.getAllUsersModel();
+    res.status(200).json(users);
+  } catch (error) {
+    res
+      .status(500)
+      .json({ message: "Error retrieving users", error: error.message });
+  }
+};
 
+// Get a user by ID
+export const getUserById = async (req, res) => {
+  try {
+    const user = await UserModel.getUserByIdModel(req.params.id);
+    if (!user) return res.status(404).json({ message: "User not found" });
+    res.status(200).json(user);
+  } catch (error) {
+    res
+      .status(500)
+      .json({ message: "Error retrieving user", error: error.message });
+  }
+};
+
+// Create a new user
+export const createUser = async (req, res) => {
+  try {
+    await UserModel.createUserModel(req.body);
+    res.status(201).json({ message: "User created successfully" });
+  } catch (error) {
+    res
+      .status(500)
+      .json({ message: "Error creating user", error: error.message });
+  }
+};
+
+// Update a user
+export const updateUser = async (req, res) => {
+  try {
+    await UserModel.updateUserModel(req.params.id, req.body);
+    res.status(200).json({ message: "User updated successfully" });
+  } catch (error) {
+    res.status(404).json({ message: "User not found", error: error.message });
+  }
+};
+
+// Delete a user
+export const deleteUser = async (req, res) => {
+  try {
+    await UserModel.deleteUserModel(req.params.id);
+    res.status(200).json({ message: "User deleted successfully" });
+  } catch (error) {
+    res
+      .status(500)
+      .json({ message: "Error deleting user", error: error.message });
+  }
+};
