--- backend/src/routes/user.routes.js
@@ -1,9 +1,0 @@
-
-import express from 'express';
-import { getUser } from '../controllers/users.controller.js';
-const router = express.Router();
-
-router.get('/get-user/id/:id', getUser);
-
-export default router;
-
